# docs/Dynasty OS-openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dynasty OS API",
    "description": "API documentation for the **Dynasty OS** application.\n\n**Authentication:** Most endpoints require the `api_key` header. If using Swagger UI, click the **Authorize** button at the top of the page and enter your API key before trying any endpoint.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://dynasty-wealth-core.base44.app/api",
      "description": "API server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/entities/Property": {
      "get": {
        "tags": [
          "Entities - Property"
        ],
        "summary": "List Property records",
        "operationId": "list_Property",
        "description": "Retrieve a list of Property records with optional filtering, sorting, and pagination.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "JSON query filter, e.g. {\"status\":\"active\"}"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            },
            "description": "Maximum number of records to return"
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Number of records to skip (pagination)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Field name to sort by. Prefix with '-' for descending order, e.g. -created_date"
          }
        ],
        "responses": {
          "200": {
            "description": "List of records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Property"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entities - Property"
        ],
        "summary": "Create a Property record",
        "operationId": "create_Property",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Property"
              },
              "example": {
                "address": "Example address",
                "status": "Active — Rehab"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Property"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - Property"
        ],
        "summary": "Delete multiple Property records",
        "operationId": "delete_many_Property",
        "description": "**Warning:** Passing an empty `{}` query will delete ALL records in this entity.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Query filter to match records to delete"
              },
              "example": {
                "address": "Example address"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "deleted": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/entities/Property/{Property_id}": {
      "get": {
        "tags": [
          "Entities - Property"
        ],
        "summary": "Get a Property record by ID",
        "operationId": "get_Property",
        "parameters": [
          {
            "name": "Property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Property"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entities - Property"
        ],
        "summary": "Update a Property record",
        "operationId": "update_Property",
        "parameters": [
          {
            "name": "Property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Property"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Property"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Entities - Property"
        ],
        "summary": "Delete a Property record",
        "operationId": "delete_Property",
        "description": "Soft-deletes the record. Use the restore endpoint to undo.",
        "parameters": [
          {
            "name": "Property_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Record ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "api_key",
        "description": "Your API key. Obtain it from your user profile or via `User.me()` in the app SDK."
      }
    },
    "schemas": {
...
      "Property": {
        "type": "object",
        "properties": {
          "property_id": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "property_type": {
            "type": "string",
            "enum": [
              "Single Family",
              "Multi Family",
              "Commercial",
              "Land",
              "Mixed Use"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "Active — Rehab",
              "Active — Rental",
              "Under Contract",
              "Listed",
              "Sold",
              "Vacant",
              "Pipeline"
            ]
          },
          "purchase_date": {
            "type": "string",
            "format": "date"
          },
          "purchase_price": {
            "type": "number"
          },
          "land_value": {
            "type": "number"
          },
          "building_value": {
            "type": "number"
          },
          "closing_costs": {
            "type": "number"
          },
          "arv": {
            "type": "number",
            "description": "After Repair Value"
          },
          "rehab_budget": {
            "type": "number"
          },
          "rehab_actual": {
            "type": "number"
          },
          "loan_balance": {
            "type": "number"
          },
          "loan_type": {
            "type": "string"
          },
          "monthly_payment": {
            "type": "number"
          },
          "monthly_rent": {
            "type": "number"
          },
          "noi": {
            "type": "number",
            "description": "Net Operating Income"
          },
          "dscr": {
            "type": "number",
            "description": "Debt Service Coverage Ratio"
          },
          "cap_rate": {
            "type": "number"
          },
          "equity": {
            "type": "number"
          },
          "exit_strategy": {
            "type": "string",
            "enum": [
              "BRRR",
              "Flip",
              "Wholesale",
              "Hold",
              "Development",
              "Land Flip"
            ]
          },
          "entity": {
            "type": "string",
            "description": "Owning LLC"
          },
          "notes": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "description": "Unique record identifier"
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record creation timestamp"
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "description": "Record last update timestamp"
          },
          "created_by_id": {
            "type": "string",
            "description": "ID of the user who created the record"
          }
        },
        "required": [
          "address",
          "status"
        ],
        "description": "Property entity"
      },
...
    }
  },
...
}